# app/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuickEx Backend API
  description: >
    QuickEx is a Stellar-based payment platform enabling payment links,
    recurring subscriptions, username resolution, marketplace trading,
    fiat on/off ramps, and webhook-driven event notifications.

    **Base URL (production):** `https://api.quickex.example.com`
    **Base URL (development):** `http://localhost:3000`

    Interactive Swagger UI is available at `/docs` when the server is running.
  version: 1.0.0
  contact:
    name: QuickEx API Support
    url: https://github.com/quickex/quickex
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.quickex.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - {}
paths:
...
  /status:
    get:
      tags: [health]
      operationId: getPublicStatus
      summary: Public status page
      description: >
        Returns a minimal, cacheable status suitable for a public status page.
        No sensitive operational details are exposed.
        Helps users distinguish client errors vs platform downtime.
      security: []
      responses:
        "200":
          description: Public status information
          headers:
            ETag:
              schema:
                type: string
              description: Cache validator for conditional requests
            Cache-Control:
              schema:
                type: string
              description: Caching directives
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStatusResponse"
        "304":
          description: Not modified (use cached version)
        "429":
          description: Too many requests (rate limited)

  # ---------------------------------------------------------------------------
  # Usernames
  # ---------------------------------------------------------------------------
...
  /api-keys:
    post:
      tags: [api-keys]
      operationId: createApiKey
      summary: Create a new API key
      description: The raw key is returned ONCE in the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created

    get:
      tags: [api-keys]
      operationId: listApiKeys
      summary: List API keys (masked)
      parameters:
        - name: owner_id
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of API keys
  /api-keys/usage:
    get:
      tags: [api-keys]
      operationId: getApiKeyUsage
      summary: Get API key usage/quota stats
      parameters:
        - name: owner_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Usage statistics
  /api-keys/{id}:
    delete:
      tags: [api-keys]
      operationId: revokeApiKey
      summary: Revoke (soft-delete) an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Key revoked
  /api-keys/{id}/rotate:
    post:
      tags: [api-keys]
      operationId: rotateApiKey
      summary: Rotate an API key
      description: Invalidates current key and issues a new one.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Key rotated

  # ---------------------------------------------------------------------------
  # Fiat Ramps
  # ---------------------------------------------------------------------------
...
  /reconciliation/trigger:
    post:
      tags: [reconciliation]
      operationId: triggerReconciliation
      summary: Manually trigger a reconciliation run
      responses:
        "200":
          description: Reconciliation completed
        "409":
          description: Already running
  /reconciliation/backfill:
    post:
      tags: [reconciliation]
      operationId: startBackfill
      summary: Trigger a backfill job for a ledger range
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackfillConfig"
      responses:
        "200":
          description: Backfill completed
        "409":
          description: Backfill already running
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  schemas:
...
    PublicStatusComponent:
      type: object
      properties:
        name:
          type: string
          example: horizon
        status:
          type: string
          enum: [operational, degraded, down]
          example: operational
        detail:
          type: string
          example: "Network: testnet"
    PublicStatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [operational, degraded, down]
          example: operational
          description: Overall platform status
        network:
          type: string
          example: testnet
          description: Stellar network name
        lastLedger:
          type: integer
          example: 12345678
          description: Last processed ledger sequence
        timestamp:
          type: string
          format: date-time
        components:
          type: array
          items:
            $ref: "#/components/schemas/PublicStatusComponent"
        version:
          type: string
          example: "0.1.0"

    # --- Usernames ---
...
    CreateApiKeyRequest:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
          maxLength: 100
        scopes:
          type: array
          items:
            type: string
            enum:
              [
                links:read,
                links:write,
                transactions:read,
                usernames:read,
                refunds:write,
                admin,
              ]
          minItems: 1
        owner_id:
          type: string

    # --- Fiat Ramps ---
...
    BackfillConfig:
      type: object
      properties:
        startLedger:
          type: integer
        endLedger:
          type: integer

    # --- Job Queue ---
...